+++ Back-End/app.js
@@ -4,9 +4,9 @@ const app = express();
 const port = process.env.APP_PORT;
 const path = require("path");
 
-// const postRouter = require("./routes/post-routes");
 const userRouter = require("./routes/user-routes");
 const authRouter = require("./routes/auth-routes");
+const messageRouter = require("./routes/message-routes");
 
 const errorAppMiddleware = require("./middleware/errorAppMiddleware");
 const { notFoundMiddleware } = require("./middleware/notFoundMiddleware");
@@ -31,6 +31,7 @@ app.use(appTimestampMiddleware);
 
 app.use(authRouter);
 app.use(authMiddleware.authRoutesMiddleware, userRouter);
+app.use(messageRouter);
 
 app.all("*", notFoundMiddleware);
