+++ Back-End/controllers/messageController.js
@@ -1,0 +1,30 @@
+const Message = require("../model/message-model");
+const http = require("http-status-codes");
+// const sequelize = require("../database/integradorConnection");
+// const { QueryTypes } = require("sequelize");
+
+// -----------------------------------------------------------------------------------------------------------------------//
+
+const receivedMessagesById = async (req, res) => {
+  //  ...
+};
+
+// -----------------------------------------------------------------------------------------------------------------------//
+
+const sentMessagesById = async (req, res) => {
+  //  ...
+};
+
+// -----------------------------------------------------------------------------------------------------------------------//
+
+const SendMessageToId = async (req, res) => {
+  //  ...
+};
+
+// -----------------------------------------------------------------------------------------------------------------------//
+
+module.exports = {
+  receivedMessagesById,
+  sentMessagesById,
+  SendMessageToId,
+};
