+++ Back-End/controllers/userController.js
@@ -1,5 +1,4 @@
 const User = require("../model/user-model");
-// const Post = require("../model/post-model-seq");
 const http = require("http-status-codes");
 // const sequelize = require("../database/integradorConnection");
 // const { QueryTypes } = require("sequelize");
@@ -14,27 +13,6 @@ const getUserAll = async (req, res) => {
 
 // -----------------------------------------------------------------------------------------------------------------------//
 
-const receivedMessagesById = async (req, res) => {
-  //  ...
-};
-
-// -----------------------------------------------------------------------------------------------------------------------//
-
-const sentMessagesById = async (req, res) => {
-  //  ...
-};
-
-// -----------------------------------------------------------------------------------------------------------------------//
-
-const SendMessageToId = async (req, res) => {
-  //  ...
-};
-
-// -----------------------------------------------------------------------------------------------------------------------//
-
 module.exports = {
   getUserAll,
-  receivedMessagesById,
-  sentMessagesById,
-  SendMessageToId,
 };
